+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Fri, 21 Apr 2023 16:14:07 GMT and should not be manually modified.
+This log was last generated on Wed, 26 Apr 2023 17:40:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 18.12.6
+
+Wed, 26 Apr 2023 17:40:07 GMT
+
+### Patches
+
+- Spinner now allows internal configuration of animation play state ([ni/nimble@e946415](https://github.com/ni/nimble/commit/e9464153249e21394b06fc5e92cc19c002cf6828))
+
+## 18.12.5
+
+Tue, 25 Apr 2023 20:41:37 GMT
+
+### Patches
+
+- Adding collapse-all-button ([ni/nimble@6c2fa63](https://github.com/ni/nimble/commit/6c2fa633203e7877536314d01d11b96866ba5376))
+
+## 18.12.4
+
+Tue, 25 Apr 2023 18:09:46 GMT
+
+### Patches
+
+- Updating row styling. ([ni/nimble@2a0427f](https://github.com/ni/nimble/commit/2a0427f9ccd40f1a1a2ce5c4014df91ab4e1b55a))
 
 ## 18.12.3